pyproject.toml
[project]
name = "hrrr-wind-experiment"
version = "0.1.0"
description = "HRRR wind data fetching and WindNinja downscaling for wildfire applications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cartopy>=0.25.0",
    "cfgrib>=0.9.15.1",
    "dask>=2026.1.2",
    "herbie-data>=2026.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "netcdf4>=1.7.4",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "py3dep>=0.19.0",
    "rioxarray>=0.21.0",
    "ruff>=0.14.14",
    "xarray>=2025.11.0",
]
